made fixes to seeds to encorporate changes from auth0 user migrations, moved discussion get request helper to utils file to make it more accessible. Made the routing more dynamic for splash and home screen to hopefully fix navigation bars
