+++ client/AppNavigator.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Container } from 'native-base';
 import { NativeRouter, Route, Switch } from 'react-router-native';
+import Splash from './src/components/splash/Splash';
 import Home from './src/views/Home';
 import Login from './src/components/auth'
 
@@ -9,7 +10,8 @@ const AppNavigator = (props) => {
         <NativeRouter>
             <Container>
                 <Switch>
-                    <Route exact path="/" component={Home} />
+                    <Route path="/home" component={Home} />
+                    <Route exact path="/" component={Splash} />
                     <Route path="/login" component={Login} />
                 </Switch>
             </Container>
